requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
pandas
